+++ Source Codes/Motors-Prototype/calculations.h
@@ -2,6 +2,6 @@
 #define CALCULATIONS_H
 
 int calculate_pulses_for_angle(float angle);
-int calculate_RPM(int pulses_elapsed);
+float calculate_RPM(int pulses_elapsed);
 
 #endif
